requirements.txt
flask
torchvision
torch
pillow